refactor(dccpeep): extract inline temporary passes
Move compiler-tagged inline-temporary spill analysis, EXX/DE/A safety classifiers, pass_inline_temp_spill_to_stack, and pass_remove_inline_temp_markers into peep_pass_inline_temp.c. The cluster is self-contained and requires no additional shared helper API.

dccpeep.c drops from about 7,900 to about 7,500 lines. Validation: strict build clean; all 316 raw inputs byte-identical in default, -Os, and -fundocumented-z80 modes; six direct fixtures pass.
